+++ src/main/kotlin/me/rafaelldi/aspire/services/AspireServiceContributor.kt
@@ -5,21 +5,15 @@ import com.intellij.execution.services.SimpleServiceViewDescriptor
 import com.intellij.openapi.project.Project
 import me.rafaelldi.aspire.AspireIcons
 import me.rafaelldi.aspire.sessionHost.AspireSessionHostManager
-import me.rafaelldi.aspire.settings.AspireSettings
 
 class AspireServiceContributor : ServiceViewContributor<AspireSessionHostServiceContributor> {
     override fun getViewDescriptor(project: Project) =
         SimpleServiceViewDescriptor("Aspire", AspireIcons.Service)
 
-    override fun getServices(project: Project): MutableList<AspireSessionHostServiceContributor> {
-        val showServices = AspireSettings.getInstance().showServices
-        return if (showServices)
-            AspireSessionHostManager.getInstance(project)
-                .getSessionHosts()
-                .toMutableList()
-        else
-            mutableListOf()
-    }
+    override fun getServices(project: Project) =
+        AspireSessionHostManager.getInstance(project)
+            .getSessionHosts()
+            .toMutableList()
 
     override fun getServiceDescriptor(project: Project, host: AspireSessionHostServiceContributor) =
         host.getViewDescriptor(project)