+++ src/main/kotlin/me/rafaelldi/aspire/settings/AspireConfigurable.kt
@@ -14,10 +14,6 @@ class AspireConfigurable: BoundConfigurable(AspireBundle.message("configurable.A
                 .bindSelected(settings::checkForNewVersions)
         }
         row {
-            checkBox(AspireBundle.message("configurable.Aspire.show.service"))
-                .bindSelected(settings::showServices)
-        }
-        row {
             checkBox(AspireBundle.message("configurable.Aspire.collect.telemetry"))
                 .bindSelected(settings::collectTelemetry)
         }