+++ src/main/kotlin/me/rafaelldi/aspire/settings/AspireSettingsState.kt
@@ -4,6 +4,5 @@ import com.intellij.openapi.components.BaseState
 
 class AspireSettingsState : BaseState() {
     var checkForNewVersions by property(true)
-    var showServices by property(false)
     var collectTelemetry by property(false)
 }