+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Update Aspire dashboard
 
+### Removed
+
+- Setting to show services
+
 ## [0.2.4] - 2024-02-13
 
 ### Added